Remove SingleTagExtension (unnecessary after all)
The only reason I couldn't get singleton/empty/... tag parsing to work
was because I was calling parse_statements(), which must be called at
the end of an opening tag and will parse up to a closing tag. If you
just don't call anything at the end of an extension opening tag parse
method, Jinja's outer parsing logic will expect the end of the opening
tag but not closing tag, exactly as intended.